Add OAuth2, JWT auth, and react-admin that uses same auth method (#21)
* add passing test for oauth2 post to token

* run black and prettier on files, add true JWT creation method

* refactor tests and put in real JWT creation method

* add build script and create initial user/password in cookiecutter template

* change sqlalchemy logo size

* fix type error on react-admin page

* add note about security

* add API docs image to README and more on security

+++ {{cookiecutter.project_slug}}/backend/app/db/crud.py
@@ -1,20 +1,9 @@
-from fastapi import HTTPException
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
-from passlib.context import CryptContext
-from fastapi.encoders import jsonable_encoder
 import typing as t
 
 from . import models, schemas
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-def get_password_hash(password: str) -> str:
-    return pwd_context.hash(password)
-
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
+from app.core.security import get_password_hash
 
 
 def get_user(db: Session, user_id: int):
@@ -24,20 +13,19 @@ def get_user(db: Session, user_id: int):
     return user
 
 
-def get_user_by_email(db: Session, email: str):
+def get_user_by_email(db: Session, email: str) -> schemas.UserBase:
     return db.query(models.User).filter(models.User.email == email).first()
 
 
-def get_users(db: Session, skip: int = 0, limit: int = 100) -> t.List[schemas.UserBase]:
+def get_users(
+    db: Session, skip: int = 0, limit: int = 100
+) -> t.List[schemas.UserBase]:
     return db.query(models.User).offset(skip).limit(limit).all()
 
 
 def create_user(db: Session, user: schemas.UserCreate):
     hashed_password = get_password_hash(user.password)
-    db_user = models.User(
-        email=user.email,
-        hashed_password=hashed_password
-    )
+    db_user = models.User(email=user.email, hashed_password=hashed_password)
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
@@ -46,12 +34,16 @@ def create_user(db: Session, user: schemas.UserCreate):
 
 def delete_user(db: Session, user_id: int):
     user = get_user(db, user_id)
+    if not user:
+        raise HTTPException(status.HTTP_404_NOT_FOUND, detail="User not found")
     db.delete(user)
     db.commit()
     return user
 
 
-def edit_user(db: Session, user_id: int, user: schemas.UserEdit) -> schemas.User:
+def edit_user(
+    db: Session, user_id: int, user: schemas.UserEdit
+) -> schemas.User:
     db_user = get_user(db, user_id)
     db_user.hashed_password = get_password_hash(user.password)
     db_user.is_active = user.is_active
